# Conflicts:
#	app/src/main/java/pl/llp/aircasting/data/api/params/MeasurementParams.kt
#	app/src/main/java/pl/llp/aircasting/data/api/repositories/ActiveSessionMeasurementsRepository.kt
#	app/src/main/java/pl/llp/aircasting/data/api/services/DownloadMeasurementsCallback.kt
#	app/src/main/java/pl/llp/aircasting/ui/view/screens/dashboard/charts/ChartRefreshService.kt

+++ app/src/main/java/pl/llp/aircasting/ui/view/screens/dashboard/charts/ChartAveragesCreator.kt
@@ -102,7 +102,7 @@ class ChartAveragesCreator {
             for (dataChunk in periodData) {
                 if (xValue > MAX_AVERAGES_AMOUNT) return entries
 
-                val yValue = getAverage(dataChunk).toDouble()
+                val yValue = getAverage(dataChunk)
                 entries.add(
                     Entry(
                         xValue.toFloat(),
